requirements.txt
streamlit==1.28.0
openai-whisper==20231117
torch==2.1.0
transformers==4.35.0
pydub==0.25.1
nltk==3.8.1
spacy==3.7.2
python-docx==1.1.0
pandas==2.1.0